p4i integration Integrate change 60859 from 3.0 to main. Description for change 60859:
: Change 60859 by vohi@vohi-hokkaido-nt on 2002/04/02 22:02:26
: 
: Fix some compile issues.
: 
: Affected files ...
: 
: ... //depot/qt/3.0/src/dialogs/qdialog.cpp#16 edit
: ... //depot/qt/3.0/src/tools/qfileinfo_win.cpp#8 edit
: ... //depot/qt/3.0/src/widgets/qspinbox.cpp#17 edit
: ... //depot/qt/3.0/src/xml/qsvgdevice.cpp#12 edit

[git-p4: depot-paths = "//depot/qt/main/": change = 60860]

+++ src/xml/qsvgdevice.cpp
@@ -1250,7 +1250,7 @@ void QSvgDevice::drawPath( const QString &data )
 	};
 	lastMode = mode;
 	// array almost full ? expand for next loop
-	if ( pcount >= path.size() - 4 )
+	if ( pcount >= (int)path.size() - 4 )
 	    path.resize( 2 * path.size() );
     }
